pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-sales-analytics"
version = "0.1.0"
description = "Production-like analytics module for AI sales bots"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "AI Sales Analytics Team" }
]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.30",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  "python-dotenv>=1.0.0",
  "pandas>=2.2.0",
  "numpy>=1.26.0",
  "matplotlib>=3.8.0",
  "seaborn>=0.13.0",
  "jinja2>=3.1.0",
  "typer>=0.12.0",
  "apscheduler>=3.10.4",
  "httpx>=0.27.0",
  "openai>=1.35.0",
  "python-dateutil>=2.9.0",
  "orjson>=3.10.0",
  "pyyaml>=6.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.5.0",
  "mypy>=1.10.0"
]

[project.scripts]
ai-sales-analytics = "ai_sales_analytics.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]
ignore = ["N818", "E501", "B008", "UP017"]
